add file for product

+++ shop/src/java/dao/CategoryDAO.java
@@ -43,7 +43,7 @@ public class CategoryDAO {
         ArrayList<Category> list = new ArrayList<>();
         while (rs.next()) {
             Category category = new Category();
-            category.setCategoryID(rs.getInt("category_id"));
+            category.setCategoryID(rs.getLong("category_id"));
             category.setCategoryName(rs.getString("category_name"));
             category.setCategoryGroup(rs.getString("category_group"));
             list.add(category);
